redis: basic integration test for redis_proxy (#6450)
