+++ test/config/utility.h
@@ -72,8 +72,6 @@ public:
   // A basic configuration (admin port, cluster_0, one listener) with no network filters.
   static const std::string BASE_CONFIG;
 
-  // A basic configuration for redis server proxying.
-  static const std::string REDIS_PROXY_CONFIG;
   // A basic configuration for L4 proxying.
   static const std::string TCP_PROXY_CONFIG;
   // A basic configuration for L7 proxying.
